change floatfield to decimalfield

+++ tracex_project/extraction/models.py
@@ -72,13 +72,16 @@ class Prompt(models.Model):
     name = models.CharField(max_length=100)
     text = models.TextField()
 
+
 class Metric(models.Model):
     """Model for metrics which are being tracked by the metrics analyzer"""
 
-    event = models.OneToOneField(Event, on_delete=models.CASCADE, related_name='metrics')
+    event = models.OneToOneField(
+        Event, on_delete=models.CASCADE, related_name="metrics"
+    )
     activity_relevance = models.CharField(max_length=25)
     timestamp_correctness = models.BooleanField()
-    correctness_confidence = models.FloatField()
+    correctness_confidence = models.DecimalField(max_digits=3, decimal_places=2)
     last_modified = models.DateTimeField(auto_now=True)
     manager = models.Manager()
